[MOB-10951] Add mobile framework info
